requirements.txt
click==7.1.2
cx-Freeze==6.3
dataclasses==0.6
future==0.18.2
importlib-metadata==2.0.0
joblib==0.17.0
nltk==3.5
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
regex==2020.10.28
six==1.15.0
torch==1.7.0+cpu
torchaudio==0.7.0
torchvision==0.8.1+cpu
tqdm==4.51.0
typing-extensions==3.7.4.3
zipp==3.4.0
